Impact about training events
Recommendation from Toby:   In "Measuring the impact of our workshops a training", adjust to refer to both Instructor Training and CLDT, i.e. "We conduct pre- and post-workshop surveys for Carpentries Workshops & Training Events." We cannot provide a preview of the pre- and post-training surveys for CLDT just yet, but I plan to migrate it into TypeForm soon.