Update host-workshop.md
Updated verbiage so that Centrally and Self-Organised workshop references link to their definitions in the respective handbook (links are not live yet, so I will need to go back and add once they are). 

Deleted information about guidelines - we will not have requirements included in the FAQ. Will link out to that instead. 

Rearranged some information for ease of reading with all of the changes.